Add files via upload
It is academic project done in 2022 using php as a backend and Html, CSS and Javascript as frontend. This is online vehicle insurance buying system here login and registration page is the staring of the website. Actually this is demo of how actual insurance system like policy bazar Actual System they uses API for vehicle details Here a demo table stores vehicle details